Extract VersionGradleFile into its own file
Move the `VersionGradleFile` parser object out of `CheckVersionIncrement.kt`
into a dedicated `VersionGradleFile.kt` in the same `io.spine.gradle.publish`
package. No behavior change.
