+++ buildSrc/src/main/kotlin/io/spine/gradle/publish/CheckVersionIncrement.kt
@@ -295,62 +295,3 @@ private fun gitShow(rootDir: File, spec: String): GitResult {
         errFile.delete()
     }
 }
-
-/**
- * Resolves the string values of `extra` properties declared in a `version.gradle.kts` file.
- *
- * Handles the shapes used across Spine repositories (see the `bump-version` skill):
- *
- *  1. a literal: `val versionToPublish: String by extra("2.0.0-SNAPSHOT.182")`;
- *  2. an alias to another `extra`: `val versionToPublish by extra(compilerVersion)` paired
- *     with `val compilerVersion: String by extra("2.0.0-SNAPSHOT.043")`;
- *  3. an alias to a plain `val`: `val versionToPublish by extra(base)` paired with
- *     `val base = "2.0.0-SNAPSHOT.043"`.
- *
- * The publishing-version property is identified by [keyForValue] using the already-resolved
- * project version as an oracle, so the specific property name (`versionToPublish`,
- * `validationVersion`, `compilerVersion`, …) does not need to be hard-coded.
- *
- * Only a single alias hop is resolved; an alias to another alias falls through to `null`,
- * which the caller treats as "publishing version not identified" and skips the check.
- */
-internal object VersionGradleFile {
-
-    private val literalExtra =
-        Regex("""val\s+(\w+)\s*(?::\s*String)?\s+by\s+extra\(\s*"([^"]+)"\s*\)""")
-    private val aliasExtra =
-        Regex("""val\s+(\w+)\s*(?::\s*String)?\s+by\s+extra\(\s*([A-Za-z_]\w*)\s*\)""")
-    private val plainAssignment =
-        Regex("""val\s+(\w+)\s*(?::\s*String)?\s*=\s*"([^"]+)"""")
-
-    /**
-     * Resolves every named `extra` (and the plain `val`s an `extra` may alias) to its
-     * string value.
-     */
-    private fun parse(content: String): Map<String, String> {
-        val literals = literalExtra.findAll(content)
-            .associate { it.groupValues[1] to it.groupValues[2] }
-        val plains = plainAssignment.findAll(content)
-            .associate { it.groupValues[1] to it.groupValues[2] }
-        val resolved = literals.toMutableMap()
-        aliasExtra.findAll(content).forEach { match ->
-            val name = match.groupValues[1]
-            val source = match.groupValues[2]
-            if (name !in resolved) {
-                (literals[source] ?: plains[source])?.let { resolved[name] = it }
-            }
-        }
-        return resolved
-    }
-
-    /**
-     * The name of the property whose resolved value equals [value], or `null` if none does.
-     */
-    fun keyForValue(content: String, value: String): String? =
-        parse(content).entries.firstOrNull { it.value == value }?.key
-
-    /**
-     * The resolved value of the property named [key], or `null` if it is absent.
-     */
-    fun valueForKey(content: String, key: String): String? = parse(content)[key]
-}
